\name{Pauli_M}
\alias{Pauli_M}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Pauli Matrices
}
\description{
The function returns the identity or the Pauli matrices. i=0 returns the identity. i=1-3, returns sigma_1, sigma_2 and sigma_3.
}
\usage{
Pauli_M(i)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{i}{
Number between 0 and 4
}
}
\details{

}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
